Adding better support for bst and cls files.
Script should now handle bst files whether or not the .bst extension is
given at the "bibliographystyle" command.

The script now also looks to see if there is a path to the
"documentclass" file (with or without .cls extension), and if it exists,
it adds it to the files to copy.